modal/requirements.txt
modal>=0.66.0
boto3>=1.35.0
biopython>=1.84
numpy>=1.26.0
scipy>=1.11.0
requests>=2.32.0
